removed credentials.json and added login route
