+++ app/server/src/controllers/users.controller.js
@@ -7,7 +7,7 @@ const getAllUsers = (req, res) => {
 };
 
 const getUser = async (req, res) => {
-  try{const user = await usersService.getUser(req.params.userId);
+  try{const user = await usersService.getUser(req);
     res.status(200).json({
       message: "Get an existing user: ",
       data: user});
@@ -45,10 +45,26 @@ const deleteUser = (req, res) => {
   res.json({ message: "Delete an existing user" });
 };
 
+const loginUser = async (req, res) => {
+  try{
+    const user = await usersService.loginUser(req.body);
+    res.status(200).json({
+      message: "Login succeded: ",
+      data: user});
+  } catch(error){
+    res.status(500).json({
+      message: "Login failed.",
+      error: error.message
+    });
+  }
+
+};
+
 module.exports = {
   getAllUsers,
   getUser,
   createUser,
   updateUser,
-  deleteUser
+  deleteUser,
+  loginUser
 };
