+++ app/server/src/database/db.json
@@ -1,1 +1,1 @@
-{"users":[{"username":"Fra","password":"prova","role":"patient","uuid":"31518575-86c5-4d25-b592-4a5118160921"}]}
+{"users":[{"username":"Fra","password":"prova","role":"patient","uuid":"31518575-86c5-4d25-b592-4a5118160921"},{"username":"Pietro","password":"prova","role":"patient","uuid":"30c688a8-95d8-4d2c-9aa3-7edc4cfc7122"}]}
